Separate tests for immutable attributes and factory
# Conflicts:
#	subprojects/core/src/test/groovy/org/gradle/api/internal/attributes/DefaultImmutableAttributesFactoryTest.groovy
